Fix group call status
